requirements.txt
chromadb==1.0.15
streamlit==1.48.0
sentence-transformers==3.0.1
pymupdf==1.24.9
pydantic==2.8.2
python-dotenv==1.0.1
pandas==2.2.2
ollama==0.3.1
tqdm==4.66.5